Day 7
I am still working on a problem. When I try flipping the image when the user presses the opposite key, the image does flip but only flips once and it resets back to its original format. I am trying to make it where it stays flipped until the opposite key is pressed.